requirements.txt
Flask==3.0.0
PyMuPDF==1.23.7
sentence-transformers==2.2.2
faiss-cpu==1.7.4
langchain==0.1.0
langchain-community==0.0.9
groq==0.4.2
python-dotenv==1.0.0
gunicorn==21.2.0
numpy==1.26.2
transformers==4.36.2
requests==2.31.0
